Move generic vertical prefix-sum scan to common/math
compute_vertical_integral is a generic top-down running sum; move it
(with its _accumulate_from_top scan) into vertical_operations.py. The
physics-weighted compute_vertical_integral_diagnostics stays in tmx and
imports the common operator.
