+++ model/atmosphere/subgrid_scale_physics/tmx/src/icon4py/model/atmosphere/subgrid_scale_physics/tmx/stencils/compute_vertical_integral_diagnostics.py
@@ -7,10 +7,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 import gt4py.next as gtx
 
-from icon4py.model.atmosphere.subgrid_scale_physics.tmx.stencils.compute_vertical_integral import (
-    _compute_vertical_integral,
-)
 from icon4py.model.common import dimension as dims, field_type_aliases as fa
+from icon4py.model.common.math.vertical_operations import _compute_vertical_integral
 from icon4py.model.common.physics.thermodynamics import _internal_energy
 from icon4py.model.common.type_alias import wpfloat
